refactor: consolidate geo utilities into src/lib/geo.ts
Move calculateGeometryBounds from lib/utils.ts, createCircleBuffer from
lib/geo/buffer.ts, and haversineDistance from DetailPanel.tsx into a
single shared geo module. Update all import sites.
