+++ src/components/map/map.tsx
@@ -4,7 +4,8 @@ import mapboxgl from 'mapbox-gl'
 import MapboxDraw from '@mapbox/mapbox-gl-draw'
 import '@mapbox/mapbox-gl-draw/dist/mapbox-gl-draw.css'
 import 'mapbox-gl/dist/mapbox-gl.css'
-import { cn, calculateGeometryBounds } from '@/lib/utils'
+import { cn } from '@/lib/utils'
+import { calculateGeometryBounds } from '@/lib/geo'
 import { createRoot } from 'react-dom/client'
 import { env } from '@/env.mjs'
