+++ src/lib/utils.ts
@@ -374,84 +374,3 @@ export function buildCityUrl(cityId: string, path: string = '', locale: string =
   // Default path-based URL structure
   return `/${locale}/${cityId}${path ? `/${path}` : ''}`;
 }
-
-type GeometryBounds = {
-  bounds: { minLng: number; maxLng: number; minLat: number; maxLat: number } | null;
-  center: [number, number];
-};
-
-/**
- * Calculates bounds and center from a GeoJSON geometry
- * @param geometry The GeoJSON geometry to process
- */
-export function calculateGeometryBounds(geometry: any): GeometryBounds {
-  const DEFAULT_RETURN: GeometryBounds = {
-    bounds: null,
-    center: [23.7275, 37.9838] // Default to Athens
-  };
-
-  if (!geometry) {
-    console.log('[Location] No geometry available, using default coordinates');
-    return DEFAULT_RETURN;
-  }
-
-  try {
-    let minLng = Infinity, maxLng = -Infinity;
-    let minLat = Infinity, maxLat = -Infinity;
-
-    // Check for supported geometry types
-    if (!['Point', 'Polygon', 'MultiPolygon', 'GeometryCollection'].includes(geometry.type)) {
-      console.warn(`[Location] Unsupported geometry type: ${geometry.type}, using default coordinates`);
-      return DEFAULT_RETURN;
-    }
-
-    const processCoordinates = (coords: number[][]) => {
-      coords.forEach(point => {
-        const [lng, lat] = point;
-        minLng = Math.min(minLng, lng);
-        maxLng = Math.max(maxLng, lng);
-        minLat = Math.min(minLat, lat);
-        maxLat = Math.max(maxLat, lat);
-      });
-    };
-
-    const processGeometry = (geom: any) => {
-      if (geom.type === 'Polygon') {
-        processCoordinates(geom.coordinates[0]);
-      } else if (geom.type === 'MultiPolygon') {
-        geom.coordinates.forEach((polygon: number[][][]) => {
-          processCoordinates(polygon[0]);
-        });
-      } else if (geom.type === 'Point') {
-        const [lng, lat] = geom.coordinates;
-        minLng = Math.min(minLng, lng);
-        maxLng = Math.max(maxLng, lng);
-        minLat = Math.min(minLat, lat);
-        maxLat = Math.max(maxLat, lat);
-      }
-    };
-
-    if (geometry.type === 'GeometryCollection') {
-      geometry.geometries.forEach(processGeometry);
-    } else {
-      processGeometry(geometry);
-    }
-
-    const bounds = {
-      minLng,
-      maxLng,
-      minLat,
-      maxLat
-    };
-
-    const center: [number, number] = [
-      (minLng + maxLng) / 2,
-      (minLat + maxLat) / 2
-    ];
-
-    return { bounds, center };
-  } catch (error) {
-    console.error('[Location] Error calculating geometry bounds:', error);
-    return DEFAULT_RETURN;
-  }
-}
